@@ -1,17 +1,22 @@
 module github.com/forhsd/asynq
 
-go 1.14
+go 1.23.0
+
+toolchain go1.24.1
 
 require (
-	github.com/golang/protobuf v1.5.2
-	github.com/google/go-cmp v0.5.6
-	github.com/google/uuid v1.2.0
-	github.com/redis/go-redis/v9 v9.0.3
+	github.com/google/go-cmp v0.7.0
+	github.com/google/uuid v1.6.0
+	github.com/redis/go-redis/v9 v9.7.3
 	github.com/robfig/cron/v3 v3.0.1
-	github.com/spf13/cast v1.3.1
-	go.uber.org/goleak v1.1.12
-	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e
-	golang.org/x/time v0.0.0-20190308202827-9d24e82272b4
-	google.golang.org/protobuf v1.26.0
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	github.com/spf13/cast v1.7.1
+	go.uber.org/goleak v1.3.0
+	golang.org/x/sys v0.31.0
+	golang.org/x/time v0.11.0
+	google.golang.org/protobuf v1.36.6
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 )
