@@ -18,8 +18,6 @@ import (
 	"github.com/redis/go-redis/v9"
 )
 
-const clientName string = "asynq"
-
 // Task represents a unit of work to be performed.
 type Task struct {
 	// typename indicates the type of task to be performed.
@@ -293,7 +291,6 @@ type RedisClientOpt struct {
 
 func (opt RedisClientOpt) MakeRedisClient() interface{} {
 	return redis.NewClient(&redis.Options{
-		ClientName:   clientName,
 		Network:      opt.Network,
 		Addr:         opt.Addr,
 		Username:     opt.Username,
@@ -318,6 +315,9 @@ type RedisFailoverClientOpt struct {
 	// Use at least three sentinels to avoid problems described in
 	// https://redis.io/topics/sentinel.
 	SentinelAddrs []string
+
+	// Redis sentinel username.
+	SentinelUsername string
 
 	// Redis sentinel password.
 	SentinelPassword string
@@ -365,9 +365,9 @@ type RedisFailoverClientOpt struct {
 
 func (opt RedisFailoverClientOpt) MakeRedisClient() interface{} {
 	return redis.NewFailoverClient(&redis.FailoverOptions{
-		ClientName:       clientName,
 		MasterName:       opt.MasterName,
 		SentinelAddrs:    opt.SentinelAddrs,
+		SentinelUsername: opt.SentinelUsername,
 		SentinelPassword: opt.SentinelPassword,
 		Username:         opt.Username,
 		Password:         opt.Password,
@@ -426,7 +426,6 @@ type RedisClusterClientOpt struct {
 
 func (opt RedisClusterClientOpt) MakeRedisClient() interface{} {
 	return redis.NewClusterClient(&redis.ClusterOptions{
-		ClientName:   clientName,
 		Addrs:        opt.Addrs,
 		MaxRedirects: opt.MaxRedirects,
 		Username:     opt.Username,
@@ -551,6 +550,3 @@ func (w *ResultWriter) Write(data []byte) (n int, err error) {
 func (w *ResultWriter) TaskID() string {
 	return w.id
 }
-
-// 返回队列ID
-func (w *ResultWriter) QueueID() string { return w.qname }
