@@ -183,15 +183,15 @@ func TestHeartbeater(t *testing.T) {
 		hb.clock = clock
 
 		// Change host and pid fields for testing purpose.
-		hb.Host = tc.host
-		hb.Pid = tc.pid
+		hb.host = tc.host
+		hb.pid = tc.pid
 
 		//===================
 		// Start Phase1
 		//===================
 
 		srvState.mu.Lock()
-		srvState.Value = SrvStateActive // simulating Server.Start
+		srvState.value = srvStateActive // simulating Server.Start
 		srvState.mu.Unlock()
 
 		var wg sync.WaitGroup
@@ -274,7 +274,7 @@ func TestHeartbeater(t *testing.T) {
 
 		// Server state change; simulating Server.Shutdown
 		srvState.mu.Lock()
-		srvState.Value = SrvStateClosed
+		srvState.value = srvStateClosed
 		srvState.mu.Unlock()
 
 		// Wait for heartbeater to write to redis
@@ -323,7 +323,7 @@ func TestHeartbeaterWithRedisDown(t *testing.T) {
 	r := rdb.NewRDB(setup(t))
 	defer r.Close()
 	testBroker := testbroker.NewTestBroker(r)
-	state := &serverState{Value: SrvStateActive}
+	state := &serverState{value: srvStateActive}
 	hb := newHeartbeater(heartbeaterParams{
 		logger:         testLogger,
 		broker:         testBroker,
